Simulation of V2G response in NetLogo
The V2G response estimation of the proposed framework is conducted by NetLogo 6.3.0 software. The NetLogo interface models the city with a 6 × 6 grid (36 district areas), roads, traffic lights, and CS. The number of EVs and the proportion of EV types can be defined based on the city data. The input parameters of agents are allowed to be selected based on statistical data, encompassing different probabilities and proportions of decision variables at the macro and micro levels. Initially, each EV is situated in a patch, that is its start position. After data processing and EV qualification confirmation, the route planning of the responsive EVs to CS is calculated, and EV agents follow traffic rules and drive along the road patch to the destination CS. The patch represents a kilometer in the real world. Each agent's properties vary with their behavior, including EV serial number, path, SOC, destination CS.
This video displays that there are 300 private EVs and 200 commercial EVs. The V2G price is 3 RMB/kWh. The yellow graph shows the simulation of SOC of EVs, the number of aggregated EVs, and the V2G support power of responsive EVs within a time range. We can click each patch or turtles to view the parameter changes of each V2G CS and EVs.